Extensive changes for Python 3 compatibility.  Rewrote proposalparser. Moved functions out of __init__ and into main.py Relative imports Tested under astroconda and iraf27, compared with /grp/jwst/ssb/sample_jwst_data/level1b
