+++ jwst/fits_generator/input_file_types.py
@@ -268,6 +268,168 @@ def get_inputfiles(input_files):
 
     return data_files
 
+def is_fitswriter(hdulist):
+    """Returns True if the file was generated by the FITSWriter
+    FITSWriter files have a HISTORY record that includes the string
+    'FITSWriter'.  Note that MIRI DHAS LVL2 files also have the same
+    HISTORY record...
+    """
+    target_string = 'FITSWriter'
+    try:
+        history = hdulist[0].header['HISTORY']
+        for card in history:
+            if card.find(target_string) != -1:
+                return True
+        return False
+    except KeyError:
+        return False
+
+def is_ncont(hdulist):
+    """Returns True if the file was generated by the NIRCAM NCont Post
+    processor.  These files have a HISTORY record that includes the string
+    'NCont Post processor'.
+    """
+    target_string = 'NCont Post processor'
+    try:
+        history = hdulist[0].header['HISTORY']
+        for card in history:
+            if card.find(target_string) != -1:
+                return True
+        return False
+    except KeyError:
+        return False
+
+def is_swts(hdulist):
+    """Returns True if the file was generated by the NIRCAM task swts2fits.
+    These files have a HISTORY record that includes the string
+    'swts2fits'.
+    """
+    target_string = 'swts2fits'
+    try:
+        history = hdulist[0].header['HISTORY']
+        for card in history:
+            if card.find(target_string) != -1:
+                return True
+    finally:
+        return False
+
+def is_miri(hdulist):
+    """Returns True if the file is MIRI data.  These files have a
+    INSTRUME keyword set to 'MIRI'."""
+
+    keywordname = 'INSTRUME'
+    target_value = 'MIRI'
+    if hdulist[0].header[keywordname].strip() == target_value:
+        return True
+    else:
+        return False
+
+def is_miri_ifu(hdulist):
+    """Returns True if the EXP_TYPE keyword is set to MIR_MRS"""
+    keywordname = 'EXP_TYPE'
+    target_value = 'MIR_MRS'
+    try:
+        if hdulist[0].header[keywordname].strip() == target_value:
+            return True
+        else:
+            return False
+    except:
+        return False
+
+def is_miri_lrs(hdulist):
+    """Returns True if the filter (FWA_POS) is P750L"""
+    keywordname = 'FWA_POS'
+    target_value = 'P750L'
+    try:
+        if hdulist[0].header[keywordname].strip() == target_value:
+            return True
+        else:
+            return False
+    except:
+        return False
+
+def is_nirspec_fm1(hdulist):
+    """Returns True if the file is NIRSpec FM1 data.  These files
+    have a DATE-OBS in the first half of 2011"""
+    instrument = hdulist[0].header['INSTRUME']
+    if instrument != 'NIRSPEC': return False
+    date = hdulist[0].header['DATE-OBS']
+    year = date[:4]
+    target_year = '2011'
+    month = int(date[5:7])
+    target_month_range = (0, 7)
+    if year == target_year and month > target_month_range[0] and month < target_month_range[1]:
+        return True
+    return False
+
+def is_nircam_fm1(hdulist):
+    """Returns True is the file is NIRCAM FM1 data.  These files
+    have a DATE-OBS after the first half of 2012"""
+    instrument = hdulist[0].header['INSTRUME']
+    if instrument != 'NIRCAM': return False
+    date = hdulist[0].header['DATE-OBS']
+    year = int(date[:4])
+    target_year = 2012.5
+    month = int(date[5:7])
+    year = year + month / 12.0
+    if year >= target_year:
+        return True
+    return False
+
+def is_nirspec_ips(hdulist):
+    """Returns True if the file is NIRSpec IPS simulation data.  These
+    file have keywords that begin with 'IPS' - I choose IPSDET as a sentinel
+    for these type of data"""
+    keywordname = 'IPSDET'
+    try:
+        if hdulist[0].header[keywordname].strip():
+            return True
+    except:
+        return False
+
+def is_tfi(hdulist):
+    """Returns True if the INSTRUME string is 'TFI'."""
+    try:
+        if hdulist[0].header['INSTRUME'].strip() == 'TFI':
+            return True
+        else:
+            return False
+    except:
+        return False
+
+def is_niriss(hdulist):
+    """Returns True if the INSTRUME string is 'NIRISS'."""
+    try:
+        if hdulist[0].header['INSTRUME'].strip() == 'NIRISS':
+            return True
+        else:
+            return False
+    except:
+        return False
+
+def is_niriss_spec_vert(hdulist):
+    """Returns True if the NIRISS spectrum is horizontal i.e. for
+    PUPIL = GR700XD or FILTER = GR150R. """
+    try:
+        if hdulist[0].header['PWCCRPUP'].strip() == 'GR700XD' or \
+                hdulist[0].header['FWCCRFIL'].strip() == 'GR150R':
+            return True
+        else:
+            return False
+    except:
+        return False
+
+def is_niriss_spec_horiz(hdulist):
+    """Returns True if the NIRISS spectrum is vertical, i.e. for
+    FILTER = GR150C. """
+    try:
+        if hdulist[0].header['FWCCRFIL'].strip() == 'GR150C':
+            return True
+        else:
+            return False
+    except:
+        return False
+
 input_file_types = [
     InputFITSFile,
     InputDataFile,
